+++ test/builtin_fixtures.test.jl
@@ -39,4 +39,30 @@ end
     @test exitcode == 0
 end
 
+
+
+@testcase "fixed_rng" begin
+
+    rand_ints = [-1, -1, -1]
+
+    tcs = Jute.collect_testobjs() do
+        @testcase "tc1" for rng in fixed_rng("abc")
+            rand_ints[1] = rand(rng, 1:1000000)
+        end
+        @testcase "tc2" for rng in fixed_rng("abc")
+            rand_ints[2] = rand(rng, 1:1000000)
+        end
+        @testcase "tc3" for rng in fixed_rng("def")
+            rand_ints[3] = rand(rng, 1:1000000)
+        end
+    end
+
+    exitcode = nested_run(tcs)
+    @test exitcode == 0
+
+    @test rand_ints[1] == rand_ints[2]
+    @test rand_ints[1] != rand_ints[3]
+end
+
+
 end