+++ docs/src/history.md
@@ -1,6 +1,11 @@
 # Version history
 
 
+## Current development version
+
+* ADDED: `fixed_rng()` local fixture.
+
+
 ## v0.2.4 (13 Feb 2020)
 
 * FIXED: bug in reporting for julia [1.3.1, 1.4.0) (an incorrect version check).